% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebaserules_functions.R
\docType{package}
\name{firebaserules_googleAuthR}
\alias{firebaserules_googleAuthR}
\alias{firebaserules_googleAuthR-package}
\title{Firebase Rules API
Creates and manages rules that determine when a Firebase Rules-enabled service should permit a request.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:50:31
filename: /Users/mark/dev/R/autoGoogleAPI/googlefirebaserulesv1.auto/R/firebaserules_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
\item https://www.googleapis.com/auth/firebase.readonly
}
}
