% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebaserules_objects.R
\name{Issue}
\alias{Issue}
\title{Issue Object}
\usage{
Issue(description = NULL, sourcePosition = NULL, severity = NULL)
}
\arguments{
\item{description}{Short error description}

\item{sourcePosition}{Position of the issue in the `Source`}

\item{severity}{The severity of the issue}
}
\value{
Issue object
}
\description{
Issue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Issues include warnings, errors, and deprecation notices.
}
