% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebaserules_objects.R
\name{ListRulesetsResponse}
\alias{ListRulesetsResponse}
\title{ListRulesetsResponse Object}
\usage{
ListRulesetsResponse(rulesets = NULL, nextPageToken = NULL)
}
\arguments{
\item{rulesets}{List of `Ruleset` instances}

\item{nextPageToken}{The pagination token to retrieve the next page of results}
}
\value{
ListRulesetsResponse object
}
\description{
ListRulesetsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for FirebaseRulesService.ListRulesets
}

