% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebaserules_functions.R
\name{projects.test}
\alias{projects.test}
\title{Test `Source` for syntactic and semantic correctness. Issues present in therules, if any, will be returned to the caller with a description, severity,and source location.The test method will typically be executed with a developer provided`Source`, but if regression testing is desired, this method may beexecuted against a `Ruleset` resource name and the `Source` will beretrieved from the persisted `Ruleset`.The following is an example of `Source` that permits users to upload imagesto a bucket bearing their user id and matching the correct metadata:_*Example*_    // Users are allowed to subscribe and unsubscribe to the blog.    service firebase.storage {      match /users/{userId}/images/{imageName} {          allow write: if userId == request.userId              && (imageName.endsWith('.png') || imageName.endsWith('.jpg'))              && resource.mimeType.startsWith('image/')      }    }}
\usage{
projects.test(TestRulesetRequest, name)
}
\arguments{
\item{TestRulesetRequest}{The \link{TestRulesetRequest} object to pass to this method}

\item{name}{Name of the project}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/firebase
\item https://www.googleapis.com/auth/firebase.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/firebase, https://www.googleapis.com/auth/firebase.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://firebase.google.com/docs/storage/security/start}{Google Documentation}

Other TestRulesetRequest functions: \code{\link{TestRulesetRequest}}
}

