% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebaserules_objects.R
\name{Release}
\alias{Release}
\title{Firebase Rules API Objects 
Creates and manages rules that determine when a Firebase Rules-enabled service should permit a request.}
\usage{
Release(name = NULL, rulesetName = NULL, createTime = NULL,
  updateTime = NULL)
}
\arguments{
\item{name}{Resource name for the `Release`}

\item{rulesetName}{Name of the `Ruleset` referred to by this `Release`}

\item{createTime}{Time the release was created}

\item{updateTime}{Time the release was updated}
}
\value{
Release object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:50:31
filename: /Users/mark/dev/R/autoGoogleAPI/googlefirebaserulesv1.auto/R/firebaserules_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Release Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`Release` is a named reference to a `Ruleset`. Once a `Release` refers to a`Ruleset`, rules-enabled services will be able to enforce the `Ruleset`.
}
