% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebaserules_objects.R
\name{Ruleset}
\alias{Ruleset}
\title{Ruleset Object}
\usage{
Ruleset(name = NULL, source = NULL, createTime = NULL)
}
\arguments{
\item{name}{Name of the `Ruleset`}

\item{source}{`Source` for the `Ruleset`}

\item{createTime}{Time the `Ruleset` was created}
}
\value{
Ruleset object
}
\description{
Ruleset Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`Ruleset` is an immutable copy of `Source` with a globally unique identifierand a creation time.
}
