% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebaserules_objects.R
\name{Source}
\alias{Source}
\title{Source Object}
\usage{
Source(files = NULL)
}
\arguments{
\item{files}{`File` set constituting the `Source` bundle}
}
\value{
Source object
}
\description{
Source Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`Source` is one or more `File` messages comprising a logical set of rules.
}
