% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebaserules_objects.R
\name{SourcePosition}
\alias{SourcePosition}
\title{SourcePosition Object}
\usage{
SourcePosition(fileName = NULL, column = NULL, line = NULL)
}
\arguments{
\item{fileName}{Name of the `File`}

\item{column}{First column on the source line associated with the source fragment}

\item{line}{Line number of the source fragment}
}
\value{
SourcePosition object
}
\description{
SourcePosition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Position in the `Source` content including its line, column number, and anindex of the `File` in the `Source` message. Used for debug purposes.
}

