% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebaserules_objects.R
\name{TestRulesetRequest}
\alias{TestRulesetRequest}
\title{TestRulesetRequest Object}
\usage{
TestRulesetRequest(source = NULL)
}
\arguments{
\item{source}{`Source` to be checked for correctness}
}
\value{
TestRulesetRequest object
}
\description{
TestRulesetRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for FirebaseRulesService.TestRuleset.
}
\seealso{
Other TestRulesetRequest functions: \code{\link{projects.test}}
}

