% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebaserules_objects.R
\name{TestRulesetResponse}
\alias{TestRulesetResponse}
\title{TestRulesetResponse Object}
\usage{
TestRulesetResponse(issues = NULL)
}
\arguments{
\item{issues}{Syntactic and semantic `Source` issues of varying severity}
}
\value{
TestRulesetResponse object
}
\description{
TestRulesetResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for FirebaseRulesService.TestRuleset.
}
