% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebaserules_objects.R
\name{File}
\alias{File}
\title{File Object}
\usage{
File(content = NULL, name = NULL, fingerprint = NULL)
}
\arguments{
\item{content}{Textual Content}

\item{name}{File name}

\item{fingerprint}{Fingerprint (e}
}
\value{
File object
}
\description{
File Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`File` containing source content.
}

