% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebaserules_objects.R
\name{Issue}
\alias{Issue}
\title{Issue Object}
\usage{
Issue(description = NULL, severity = NULL, sourcePosition = NULL)
}
\arguments{
\item{description}{Short error description}

\item{severity}{The severity of the issue}

\item{sourcePosition}{Position of the issue in the `Source`}
}
\value{
Issue object
}
\description{
Issue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Issues include warnings, errors, and deprecation notices.
}

