% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firebaserules_objects.R
\name{ListReleasesResponse}
\alias{ListReleasesResponse}
\title{ListReleasesResponse Object}
\usage{
ListReleasesResponse(nextPageToken = NULL, releases = NULL)
}
\arguments{
\item{nextPageToken}{The pagination token to retrieve the next page of results}

\item{releases}{List of `Release` instances}
}
\value{
ListReleasesResponse object
}
\description{
ListReleasesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response for FirebaseRulesService.ListReleases.
}

